+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-3549] Add new quality rule result type: "deviation"
+- [TD-2635] Admin can manually delete structures and all its children
 - [TD-3917] `PATCH /api/systems/:external_id/metadata` allows mutable metadata
   to be replaced or merged